requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
docopt==0.6.2
filelock==3.12.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
platformdirs==3.8.0
PyJWT==2.8.0
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
virtualenv==20.23.1
Werkzeug==2.3.6
yarg==0.1.9
yarl==1.9.2
